tests: extract unfinalized-upgrade driver into a mixin
Factor the version/upgrade/finalize driving primitives out of
ManualFinalizationUpgradeTest into a reusable UnfinalizedUpgradeMixin:
roll the binaries forward with features_auto_finalization=false so the
active (downgrade-floor) version is held back, observe/finalize via the
admin v2 RPCs, and roll back before finalizing.

Behavior is unchanged -- the methods move verbatim and the test inherits
them. The mixin assumes self.redpanda/installer/admin/admin_v2, so a
second test can drive the same flow against a different topology (the
target cluster of a shadow link, with a fixed source alongside). The
single-cluster-boot helper _start_at_old stays on the test since it calls
redpanda.start() directly.
